Passage du système de partition vers un référentiel relatif

+++ SharedLibs/Interfaces/IDrawable.hpp
@@ -5,8 +5,8 @@
 #ifndef R_TYPE_IDRAWABLE_HPP
 #define R_TYPE_IDRAWABLE_HPP
 
-#include <SFML/Graphics.hpp>
 #include <vec2.hpp>
+#include <SFML/Graphics.hpp>
 
 class IDrawable {
 public:
@@ -14,8 +14,8 @@ public:
 
 public:
     virtual void Draw(sf::RenderTexture &rect) = 0;
-    virtual vec2<int> GetRenderRect() = 0;
-    virtual vec2<int> GetPosition() = 0;
+    virtual vec2<float> GetRenderRect() = 0;
+    virtual vec2<float> GetPosition() = 0;
 };
 
 #endif //R_TYPE_IDRAWABLE_HPP
